revert: withdraw Bun component reporter change
